+++ src/Resource/IdentificationInspection.php
@@ -9,4 +9,6 @@ use DigitalCz\DigiSign\Resource\Traits\EntityResourceTrait;
 class IdentificationInspection extends BaseResource
 {
     use EntityResourceTrait;
+
+    public InspectionClaims $claims;
 }